Centered model and rotation controls
When a model is loaded in the viewer, it will appear centered in the
view. The controls of rotation are done around of each of the axis.
Three circles has been added in order to help in the process of
rotation. The controls are:
W/S -> rotate around x
A/D-> rotate around y
Q/E -> rotate around z
R/F -> zoom in/out
C -> show/hide guidance circles