GLUI viewer
Change on Viewer controls:
The rotation, translation and zoom is done using the interface provided
by glui. For the last two, it can be also done using the mouse and the
wheel.
The Feature class is now a simple struct.
A method has been created in pca to center the mean model everytime is
loaded.