Modify main, alphas minimized and noisy pca with normal mean
main.cpp has been cleaned so now it is only required to select the
dataset to use and the task to run.
In the viewer, now the alphas rollout will be initialized minimized so
the movement controls will be seen.
A pca model  has been generated by mixing the pca of the noisy data and
the mean of the normal pca